job-market-assistant/backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
google-generativeai==0.3.0
pinecone-client==2.2.2
requests==2.28.2
python-multipart==0.0.6
pydantic==1.10.7